backend/scraper/requirements.txt
selenium==4.17.2
webdriver-manager==4.0.1
supabase==2.3.0
python-dotenv==1.0.1
requests==2.31.0
schedule==1.2.1
gunicorn==21.2.0
